changing default implementation into macros
