+++ expander_compiler/src/zkcuda/proving_system.rs
@@ -11,6 +11,3 @@ mod expander_gkr;
 pub use expander_gkr::*;
 
 pub mod expander_gkr_parallelized;
-
-pub mod client;
-pub mod server;
