+++ expander_compiler/src/zkcuda/proving_system/expander_gkr_parallelized/client.rs
@@ -2,6 +2,18 @@ use super::server::{RequestType, SERVER_URL};
 
 use reqwest::Client;
 
+pub async fn request_setup(setup_file: &str) {
+    post_request(RequestType::Setup(setup_file.to_string())).await;
+}
+
+pub async fn request_prove() {
+    post_request(RequestType::Prove).await;
+}
+
+pub async fn request_exit() {
+    post_request(RequestType::Exit).await;
+}
+
 pub async fn post_request(request: RequestType) {
     let client = Client::new();
     let server_url = SERVER_URL;
@@ -18,16 +30,4 @@ pub async fn post_request(request: RequestType) {
     } else {
         eprintln!("Request failed: {}", res.status());
     }
-}
-
-pub async fn request_setup(setup_file: &str) {
-    post_request(RequestType::Setup(setup_file.to_string())).await;
-}
-
-pub async fn request_prove() {
-    post_request(RequestType::Prove).await;
-}
-
-pub async fn request_exit() {
-    post_request(RequestType::Exit).await;
 }
