+++ expander_compiler/src/zkcuda/proving_system/expander_gkr_parallelized/server.rs
@@ -4,7 +4,7 @@ pub static SERVER_URL: &str = "http://127.0.0.1:3000/";
 
 #[derive(Serialize, Deserialize)]
 pub enum RequestType {
-    Setup(String),       // The path to the computation graph setup file
-    Prove, // (Parallelization count, kernel_id)
+    Setup(String), // The path to the computation graph setup file
+    Prove,         // (Parallelization count, kernel_id)
     Exit,
 }
