+++ expander_compiler/src/zkcuda/proving_system/expander_gkr_parallelized/shared_memory_utils.rs
@@ -1,19 +1,17 @@
 #![allow(static_mut_refs)]
 
-use std::process::Command;
-
 use crate::{
-    circuit::layered::Circuit, frontend::SIMDField, zkcuda::{kernel::LayeredCircuitInputVec, proving_system::{CombinedProof, ExpanderGKRProvingSystem, ParallelizedExpanderGKRProvingSystem, ProvingSystem}},
+    frontend::SIMDField,
+    zkcuda::proving_system::{CombinedProof, ExpanderGKRProvingSystem},
 };
-use arith::Field;
-use gkr_engine::{ExpanderPCS, FieldEngine, FieldType, GKREngine, PolynomialCommitmentType};
+use arith::{Field, SimdField};
+use gkr_engine::{ExpanderPCS, FieldEngine, GKREngine};
 use serdes::ExpSerde;
 use shared_memory::{Shmem, ShmemConf};
 
-use crate::circuit::{config::Config, layered::InputType};
+use crate::circuit::config::Config;
 
-use super::{
-    ExpanderGKRCommitment, ExpanderGKRCommitmentExtraInfo, ExpanderGKRProof,
+use crate::zkcuda::proving_system::expander_gkr::{
     ExpanderGKRProverSetup, ExpanderGKRVerifierSetup,
 };
 
@@ -36,7 +34,11 @@ pub struct SharedMemoryEngine {}
 impl SharedMemoryEngine {
     /// Allocate shared memory for the given name and size if it is not already allocated or if the existing allocation is smaller than the target size.
     /// The result is stored in the provided `handle`, it's the caller's responsibility to ensure that the `handle` lives long enough for the reader to access the shared memory.
-    fn allocate_shared_memory_if_necessary(handle: &mut Option<Shmem>, name: &str, target_size: usize) {
+    fn allocate_shared_memory_if_necessary(
+        handle: &mut Option<Shmem>,
+        name: &str,
+        target_size: usize,
+    ) {
         if handle.is_some() && handle.as_ref().unwrap().len() >= target_size {
             return;
         }
@@ -49,7 +51,7 @@ impl SharedMemoryEngine {
                 .create()
                 .unwrap(),
         );
-    }    
+    }
 
     /// Write an object to shared memory. If the shared memory is not allocated or is too small, it will be allocated with the size of the serialized object.
     fn write_object_to_shared_memory<T: ExpSerde>(
@@ -74,10 +76,14 @@ impl SharedMemoryEngine {
         shared_memory_ref: &str,
         offset: usize,
     ) -> T {
-        let shmem = ShmemConf::new().flink(shared_memory_ref).open().expect("Failed to open shared memory");
+        let shmem = ShmemConf::new()
+            .flink(shared_memory_ref)
+            .open()
+            .expect("Failed to open shared memory");
         let object_ptr = shmem.as_ptr() as *const u8;
         let object_len = shmem.len();
-        let buffer = unsafe { std::slice::from_raw_parts(object_ptr.add(offset), object_len - offset) };
+        let buffer =
+            unsafe { std::slice::from_raw_parts(object_ptr.add(offset), object_len - offset) };
         T::deserialize_from(buffer).expect("Failed to deserialize object")
     }
 }
@@ -89,7 +95,7 @@ impl SharedMemoryEngine {
         F: FieldEngine,
         PCS: ExpanderPCS<F, PCSField>,
     >(
-    pcs_setup: &(
+        pcs_setup: &(
             ExpanderGKRProverSetup<PCSField, F, PCS>,
             ExpanderGKRVerifierSetup<PCSField, F, PCS>,
         ),
@@ -112,14 +118,87 @@ impl SharedMemoryEngine {
         Self::read_object_from_shared_memory("pcs_setup", 0)
     }
 
-    pub fn write_proof_to_shared_memory<C: Config>(proof: &CombinedProof<C, ExpanderGKRProvingSystem<C>>)
-    where 
-        C::FieldConfig: FieldEngine<SimdCircuitField = C::PCSField>,
-    {
-        Self::write_object_to_shared_memory(proof, unsafe { &mut SHARED_MEMORY.proof },"proof");
+    pub fn write_witness_to_shared_memory<F: FieldEngine>(
+        values: &[impl AsRef<[F::SimdCircuitField]>],
+    ) {
+        let total_size = std::mem::size_of::<usize>()
+            + values
+                .iter()
+                .map(|v| std::mem::size_of::<usize>() + std::mem::size_of_val(v.as_ref()))
+                .sum::<usize>();
+
+        unsafe {
+            Self::allocate_shared_memory_if_necessary(
+                &mut SHARED_MEMORY.witness,
+                "witness",
+                total_size,
+            );
+
+            let mut ptr = SHARED_MEMORY.witness.as_mut().unwrap().as_ptr();
+
+            // Copy the length of the vector
+            let len = values.len();
+            let len_ptr = &len as *const usize as *const u8;
+            std::ptr::copy_nonoverlapping(len_ptr, ptr, std::mem::size_of::<usize>());
+            ptr = ptr.add(std::mem::size_of::<usize>());
+
+            for vals in values {
+                let vals_len = vals.as_ref().len();
+                let len_ptr = &vals_len as *const usize as *const u8;
+                std::ptr::copy_nonoverlapping(len_ptr, ptr, std::mem::size_of::<usize>());
+                ptr = ptr.add(std::mem::size_of::<usize>());
+
+                let vals_size = std::mem::size_of_val(vals.as_ref());
+                std::ptr::copy_nonoverlapping(vals.as_ref().as_ptr() as *const u8, ptr, vals_size);
+                ptr = ptr.add(vals_size);
+            }
+        }
     }
 
-    pub fn read_proof_from_shared_memory<C: Config>() -> CombinedProof<C, ExpanderGKRProvingSystem<C>> 
+    pub fn read_witness_from_shared_memory<F: FieldEngine>() -> Vec<Vec<F::SimdCircuitField>> {
+        let shmem = ShmemConf::new().flink("witness").open().unwrap();
+        let mut ptr = shmem.as_ptr();
+        let n_components: usize =
+            usize::deserialize_from(unsafe { std::slice::from_raw_parts(ptr, size_of::<usize>()) })
+                .unwrap();
+        ptr = unsafe { ptr.add(size_of::<usize>()) };
+
+        (0..n_components)
+            .map(|_| {
+                let total_len_i: usize = usize::deserialize_from(unsafe {
+                    std::slice::from_raw_parts(ptr, size_of::<usize>())
+                })
+                .unwrap();
+                ptr = unsafe { ptr.add(size_of::<usize>()) };
+
+                let mut vals = Vec::with_capacity(total_len_i);
+                unsafe {
+                    std::ptr::copy_nonoverlapping(
+                        ptr as *const F::SimdCircuitField,
+                        vals.as_mut_ptr(),
+                        total_len_i,
+                    );
+                    vals.set_len(total_len_i);
+                }
+
+                ptr = unsafe { ptr.add(total_len_i * <F::SimdCircuitField as Field>::SIZE) };
+                vals
+            })
+            .collect()
+    }
+
+    pub fn write_proof_to_shared_memory<C: Config>(
+        proof: &CombinedProof<C, ExpanderGKRProvingSystem<C>>,
+    ) where
+        C::FieldConfig: FieldEngine<SimdCircuitField = C::PCSField>,
+    {
+        Self::write_object_to_shared_memory(proof, unsafe { &mut SHARED_MEMORY.proof }, "proof");
+    }
+
+    pub fn read_proof_from_shared_memory<
+        C: GKREngine,
+        ECCConfig: Config<FieldConfig = C::FieldConfig>,
+    >() -> CombinedProof<ECCConfig, ExpanderGKRProvingSystem<C>>
     where
         C::FieldConfig: FieldEngine<SimdCircuitField = C::PCSField>,
     {
